🚀 2차 배포

+++ specific_mbti_prediction.py
@@ -22,7 +22,7 @@ def mbti_train_and_prediction(mbti_type: str, answer: str):
     
     vectorizer = TfidfVectorizer()
     X_train_tfidf = vectorizer.fit_transform(X)
-    joblib.dump(vectorizer, './models/vectorizer_detail_text.sav')
+    joblib.dump(vectorizer, './models/vectorizer_detail_text.joblib')
 
     X_test = answer
 
@@ -45,7 +45,7 @@ def get_specific_mbti(mbti_type: str, answer:str):
     return result
 
 def make_detail_feedback_df(user_feedback):
-    mbti_name = {"I": "IE", "E": "IE", "S": "SN", "N": "SN", "T": "TF", "F": "TF", "P": "PJ", "J":"PJ"}
+    mbti_name = {"I": "IE", "E": "IE", "S": "SN", "N": "SN", "T": "FT", "F": "FT", "P": "PJ", "J":"PJ"}
     
     for data in user_feedback:
         mbti = data.detail_mbti
@@ -57,25 +57,30 @@ def make_detail_feedback_df(user_feedback):
             feedback_df = pd.read_csv('./feedback/detail/detail_feedback_{0}.csv'.format(mbti_name[mbti]))
             feedback_df = pd.concat([feedback_df, pd.DataFrame({'word': [answer], 'mbti': mbti})], ignore_index=True)
             
-    feedback_df.to_csv('./feedback/detail/detail_feedback_{0}.csv'.format(mbti_name[mbti]))
+        feedback_df.to_csv('./feedback/detail/detail_feedback_{0}.csv'.format(mbti_name[mbti]))
 
-def extra_train_specific_model():
-    mbti_type = ['IE', 'SN', 'TF', 'PJ']
+def extra_train_specific_model(mbti_type: str):
     
-    for m in mbti_type:
-        feedback_df = pd.read_csv('./feedback/detail/detail_feedback_{0}.csv'.format(m))
-    
+    if os.path.exists('./feedback/detail/detail_feedback_{0}.csv'.format(mbti_type)):
+        feedback_df = pd.read_csv('./feedback/detail/detail_feedback_{0}.csv'.format(mbti_type))
+        
         if len(feedback_df) >= 3:
             X = feedback_df['word']
             y = feedback_df['mbti']
         
-            vectorizer = joblib.load('./models/vectorizer_detail_text.sav')
-            clf = pickle.load(open('./models/model_detail_mbti_{0}.sav'.format(m), 'rb'))
+            vectorizer = joblib.load('./models/vectorizer_detail_text.joblib')
+            clf = pickle.load(open('./models/model_detail_mbti_{0}.sav'.format(mbti_type), 'rb'))
         
             X_train_tfidf = vectorizer.fit_transform(X)
             clf.fit(X_train_tfidf, y)
         
             joblib.dump(vectorizer, './models/vectorizer_detail_text.joblib')
-            pickle.dump(clf, open('./models/model_detail_mbti_{0}.sav'.format(m)), 'wb')
+            pickle.dump(clf, open('./models/model_detail_mbti_{0}.sav'.format(mbti_type), 'wb'))
+            
+            return True, len(feedback_df)
         else:
-            return False
+            return False, len(feedback_df)
+    
+    else:
+        return False, 0
+
